Bumping taikungoclient and fixing build

+++ cmd/backupcredential/list/list.go
@@ -2,6 +2,7 @@ package list
 
 import (
 	"context"
+
 	"github.com/itera-io/taikun-cli/api"
 	"github.com/itera-io/taikun-cli/cmd/cmdutils"
 	"github.com/itera-io/taikun-cli/config"
@@ -105,7 +106,7 @@ func listRun(opts *ListOptions) (err error) {
 			break
 		}
 
-		if backupCredentialsCount == data.GetTotalCount() {
+		if int64(backupCredentialsCount) == data.GetTotalCount() { // casting is safe, extending maximum bounds
 			break
 		}
