+++ cmd/organization/list/list.go
@@ -2,6 +2,7 @@ package list
 
 import (
 	"context"
+
 	"github.com/itera-io/taikun-cli/api"
 	"github.com/itera-io/taikun-cli/cmd/cmdutils"
 	"github.com/itera-io/taikun-cli/config"
@@ -136,7 +137,7 @@ func ListOrganizations(opts *ListOptions) (organizations []taikuncore.Organizati
 			break
 		}
 
-		if organizationsCount == data.GetTotalCount() {
+		if int64(organizationsCount) == data.GetTotalCount() { // casting is safe, extending maximum bounds
 			break
 		}
